+++ bbs/readmail.cpp
@@ -345,7 +345,7 @@ void readmail(int mode) {
   mailrec m, m1;
   char ch;
   int num_mail, num_mail1;
-  int user_number, system_number;
+  uint16_t user_number, system_number;
   net_system_list_rec *csne;
   filestatusrec fsr;
   bool attach_exists = false;
